// src/workflowComponents/enums/workflowNodeType.ts
export enum WorkflowNodeType {
  START = 0,
  PHONE_CALL = 1,
  TEXT_MESSAGE = 2,
  INTERNET = 3,
  CHAT_MESSAGE = 4,
  DELAY = 5,
  SEND_EMAIL = 6,
  SEND_SMS = 7,
  SCHEDULED_ACTIONS = 8,
  REAL_SCHEDULE = 9,
  CATCH_ALL_SCHEDULE = 10,
  SEND_TO_ROUTING_POLICY = 11,
  NO_KEYS_PRESSED_ON_IVR = 12,
  IVR = 13,
  PLAY_MESSAGE = 14,
  KEY_PRESS = 15,
  HANG_UP = 16
}